+++ packages/spacecat-shared-drs-client/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-drs-client-v1.12.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-drs-client-v1.11.1...@adobe/spacecat-shared-drs-client-v1.12.0) (2026-06-11)
+
+
+### Features
+
+* **drs-client:** forward imsOrgId/brand as parameters.metadata on sc… ([#1664](https://github.com/adobe/spacecat-shared/issues/1664)) ([8ea4a7f](https://github.com/adobe/spacecat-shared/commit/8ea4a7fd25038508198bd3711ff7a87f69b8b477))
+
 # [@adobe/spacecat-shared-drs-client-v1.11.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-drs-client-v1.11.0...@adobe/spacecat-shared-drs-client-v1.11.1) (2026-05-29)